requirements.txt
vanna[sqlite,gemini,fastapi]>=2.0.0
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
plotly
pandas